+++ summoner-cli/test/Test/TomlSpec.hs
@@ -33,7 +33,7 @@ tomlConfigSpec = describe "TOML configuration spec" $ do
     it "parses default configuration" $
         decode configCodec defaultConfigFileContent `shouldSatisfy` isRight
     it "default configuration is up-to-date" $
-        readFileText "examples/summoner-default.toml"
+        decodeUtf8 <$> readFileBS "examples/summoner-default.toml"
             `shouldReturn` defaultConfigFileContent
 
 tomlProp :: Spec